bookstore-microservice/api-gateway/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
requests>=2.31.0
PyJWT>=2.8.0
